+++ Sources/JXLTool/JXLTool.swift
@@ -30,7 +30,7 @@ struct JXLTool: ParsableCommand {
     )
 }
 
-let JXLToolVersion = "0.13.0-dev"
+let JXLToolVersion = "0.13.0"
 
 enum JXLExitCode: Int32, Error {
     case generalError = 1